requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
chardet==3.0.4
decorator==4.0.11
discord.py==0.16.8
multidict==3.1.0
pbr==3.1.1
psycopg2==2.7.1
six==1.10.0
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
websockets==3.3
